update rootpage and rowid in test_cdc

+++ tests/integration/functions/test_cdc.rs
@@ -955,13 +955,13 @@ fn test_cdc_schema_changes() {
                 Value::Null,
                 Value::Integer(1),
                 Value::Text("sqlite_schema".to_string()),
-                Value::Integer(2),
+                Value::Integer(3),
                 Value::Null,
                 Value::Blob(record([
                     Value::Text("table".to_string()),
                     Value::Text("t".to_string()),
                     Value::Text("t".to_string()),
-                    Value::Integer(3),
+                    Value::Integer(4),
                     Value::Text(
                         "CREATE TABLE t (x, y, z UNIQUE, q, PRIMARY KEY (x, y))".to_string()
                     )
@@ -973,13 +973,13 @@ fn test_cdc_schema_changes() {
                 Value::Null,
                 Value::Integer(1),
                 Value::Text("sqlite_schema".to_string()),
-                Value::Integer(5),
+                Value::Integer(6),
                 Value::Null,
                 Value::Blob(record([
                     Value::Text("table".to_string()),
                     Value::Text("q".to_string()),
                     Value::Text("q".to_string()),
-                    Value::Integer(6),
+                    Value::Integer(7),
                     Value::Text("CREATE TABLE q (a, b, c)".to_string())
                 ])),
                 Value::Null,
@@ -989,13 +989,13 @@ fn test_cdc_schema_changes() {
                 Value::Null,
                 Value::Integer(1),
                 Value::Text("sqlite_schema".to_string()),
-                Value::Integer(6),
+                Value::Integer(7),
                 Value::Null,
                 Value::Blob(record([
                     Value::Text("index".to_string()),
                     Value::Text("t_q".to_string()),
                     Value::Text("t".to_string()),
-                    Value::Integer(7),
+                    Value::Integer(8),
                     Value::Text("CREATE INDEX t_q ON t (q)".to_string())
                 ])),
                 Value::Null,
@@ -1005,13 +1005,13 @@ fn test_cdc_schema_changes() {
                 Value::Null,
                 Value::Integer(1),
                 Value::Text("sqlite_schema".to_string()),
-                Value::Integer(7),
+                Value::Integer(8),
                 Value::Null,
                 Value::Blob(record([
                     Value::Text("index".to_string()),
                     Value::Text("q_abc".to_string()),
                     Value::Text("q".to_string()),
-                    Value::Integer(8),
+                    Value::Integer(9),
                     Value::Text("CREATE INDEX q_abc ON q (a, b, c)".to_string())
                 ])),
                 Value::Null,
@@ -1021,12 +1021,12 @@ fn test_cdc_schema_changes() {
                 Value::Null,
                 Value::Integer(-1),
                 Value::Text("sqlite_schema".to_string()),
-                Value::Integer(2),
+                Value::Integer(3),
                 Value::Blob(record([
                     Value::Text("table".to_string()),
                     Value::Text("t".to_string()),
                     Value::Text("t".to_string()),
-                    Value::Integer(3),
+                    Value::Integer(4),
                     Value::Text(
                         "CREATE TABLE t (x, y, z UNIQUE, q, PRIMARY KEY (x, y))".to_string()
                     )
@@ -1039,12 +1039,12 @@ fn test_cdc_schema_changes() {
                 Value::Null,
                 Value::Integer(-1),
                 Value::Text("sqlite_schema".to_string()),
-                Value::Integer(7),
+                Value::Integer(8),
                 Value::Blob(record([
                     Value::Text("index".to_string()),
                     Value::Text("q_abc".to_string()),
                     Value::Text("q".to_string()),
-                    Value::Integer(8),
+                    Value::Integer(9),
                     Value::Text("CREATE INDEX q_abc ON q (a, b, c)".to_string())
                 ])),
                 Value::Null,
@@ -1074,13 +1074,13 @@ fn test_cdc_schema_changes_alter_table() {
                 Value::Null,
                 Value::Integer(1),
                 Value::Text("sqlite_schema".to_string()),
-                Value::Integer(2),
+                Value::Integer(3),
                 Value::Null,
                 Value::Blob(record([
                     Value::Text("table".to_string()),
                     Value::Text("t".to_string()),
                     Value::Text("t".to_string()),
-                    Value::Integer(3),
+                    Value::Integer(4),
                     Value::Text(
                         "CREATE TABLE t (x, y, z UNIQUE, q, PRIMARY KEY (x, y))".to_string()
                     )
@@ -1092,12 +1092,12 @@ fn test_cdc_schema_changes_alter_table() {
                 Value::Null,
                 Value::Integer(0),
                 Value::Text("sqlite_schema".to_string()),
-                Value::Integer(2),
+                Value::Integer(3),
                 Value::Blob(record([
                     Value::Text("table".to_string()),
                     Value::Text("t".to_string()),
                     Value::Text("t".to_string()),
-                    Value::Integer(3),
+                    Value::Integer(4),
                     Value::Text(
                         "CREATE TABLE t (x, y, z UNIQUE, q, PRIMARY KEY (x, y))".to_string()
                     )
@@ -1106,7 +1106,7 @@ fn test_cdc_schema_changes_alter_table() {
                     Value::Text("table".to_string()),
                     Value::Text("t".to_string()),
                     Value::Text("t".to_string()),
-                    Value::Integer(3),
+                    Value::Integer(4),
                     Value::Text(
                         "CREATE TABLE t (x PRIMARY KEY, y PRIMARY KEY, z UNIQUE)".to_string()
                     )
@@ -1129,12 +1129,12 @@ fn test_cdc_schema_changes_alter_table() {
                 Value::Null,
                 Value::Integer(0),
                 Value::Text("sqlite_schema".to_string()),
-                Value::Integer(2),
+                Value::Integer(3), 
                 Value::Blob(record([
                     Value::Text("table".to_string()),
                     Value::Text("t".to_string()),
                     Value::Text("t".to_string()),
-                    Value::Integer(3),
+                    Value::Integer(4),
                     Value::Text(
                         "CREATE TABLE t (x PRIMARY KEY, y PRIMARY KEY, z UNIQUE)".to_string()
                     )
@@ -1143,7 +1143,7 @@ fn test_cdc_schema_changes_alter_table() {
                     Value::Text("table".to_string()),
                     Value::Text("t".to_string()),
                     Value::Text("t".to_string()),
-                    Value::Integer(3),
+                    Value::Integer(4),
                     Value::Text(
                         "CREATE TABLE t (x PRIMARY KEY, y PRIMARY KEY, z UNIQUE, t)".to_string()
                     )
